// RMS-BackOfficePortal/src/app/core/models/fee-group.interface.ts
export interface FeeGroup {
    fee_grp_id: number;
    ss_cd: string;
    ss_fee_grp_id: number;
    fee_grp_nm_en: string;
    fee_grp_nm_bm: string;
    dtCreated: Date;
    dtModified: Date;
    createdBy: string;
    modifiedBy: string;
    status_en: string;
    status_bm: string;
    total: number;
  }
  

  